requirements.txt
# local package
-e .

# dev requirements
pytest
black
ruff

# external requirements
click
Sphinx
coverage
# awscli
flake8
python-dotenv>=0.5.1
numpy
joblib

# data science
pandas
scikit-learn
torch
tqdm
matplotlib
wandb
umap-learn

# geo
rasterio